Followed Vulkan Tutorial up until just before the graphics pipeline section.
At this point, this repo is me learning  Vulkan. When I have enough rendering capabilities, as in I can render textured models using multiple threads and to multiple windows, this will be merged with Visualizer4D. Visualizer4D will then be renamed to GBC-1.4, and worked on until I can use it for both Visualizer4D and MinecraftRecoded. I don't have a favorite of the two, so which I'll work on first is up in the air. Either way, the ideologies of GBC-1.3 were to be a game engine like Unity or Unreal. I have had enough of that type of programming for a while. I just want to have the backend capabilities and the game frontend. I don't want a middle layer of abstraction that Unity and Unreal provide. At least not until I can actually finish a project so I have SOMETHING to show for. This is going to be THE central engine architecture I use for pretty much every graphical application I intend to make for the foreseeable future. After I have it in a workable state is when I'll start working on the project's I've abandoned because I kept getting tired of the engine middle ground work. Projects like Visualizer4D, MinecraftRecoded, MinecraftSkinEditor, NDChess, EZ80IDE, NBodySimulator.